micro/customer_service/requirements.txt
Django==5.2.10
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
mysqlclient==2.2.6
django-cors-headers==4.6.0
